requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-crispy-forms==1.14.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
Pygments==2.11.2
pytz==2021.3
requests==2.29.0
sqlparse==0.4.2
urllib3==1.26.15
